added KinematicBody, framerate capping, and delta time

+++ src/main/kotlin/io/jcurtis/kadot/TestNode.kt
@@ -3,8 +3,8 @@ package io.jcurtis.kadot
 import io.jcurtis.kadot.engine.nodes.Node
 import io.jcurtis.kadot.engine.nodes.NodeType
 
-class TestNode: Node(NodeType.NODE, "Test Node") {
-    override fun update(delta: Long) {
+class TestNode : Node(NodeType.NODE, "Test Node") {
+    override fun update(delta: Double) {
         println("Test Node Update")
     }
 }
